Delete scribe.exe
The .exe was being embellished with something dangerous that I didn't intend. I think this has to do with the security risks involved with hosting an .exe file online. So, I have decided to remove that from the repository completely. You will have to compile this program yourself for safety purposes.